# backend/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Session struct {
	ID        pgtype.UUID        `json:"id"`
	StartedAt pgtype.Timestamptz `json:"started_at"`
}

type Turn struct {
	ID        pgtype.UUID        `json:"id"`
	SessionID pgtype.UUID        `json:"session_id"`
	Role      string             `json:"role"`
	Language  string             `json:"language"`
	Text      string             `json:"text"`
	LatencyMs *int32             `json:"latency_ms"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}